refactor: split summarize.py (961 lines) into 5 focused modules
Split the Python summarize worker into:
- summarize_text.py (98) — text normalization + segmentation
- summarize_extract.py (371) — regex patterns + deterministic extraction
- summarize_llm.py (93) — LLM fallback decision classification
- summarize.py (256) — orchestration + single-transcript entry
- summarize_batch.py (193) — batch processing + stdin/stdout CLI

80/80 existing tests pass. No TypeScript changes needed.

Codex adversarial review: 4/5 PASS
