+++ cloud/workers/summarize.py
@@ -1,138 +1,49 @@
 #!/usr/bin/env python3
-"""
-Summarize Worker - Generates decision summaries for probe transcripts.
-
-Protocol:
-- Reads JSON input from stdin
-- Writes JSON output to stdout
-- Logs structured JSON to stderr
-
-Input format (SummarizeWorkerInput):
-{
-  "transcriptId": string,
-  "modelId": string,           // Summary LLM model to use
-  "transcriptContent": {       // The transcript content from probe
-    "turns": [
-      {
-        "probePrompt": string,
-        "targetResponse": string,
-        ...
-      }
-    ]
-  }
-}
-
-Output format (SummarizeWorkerOutput):
-Success:
-{
-  "success": true,
-  "summary": {
-    "decisionCode": string,    // Positive integer string, "refusal", or "other"
-    "decisionSource": string,  // "deterministic" or "llm"
-    "decisionText": string     // 50-word explanation
-  }
-}
-
-Error:
-{
-  "success": false,
-  "error": {
-    "message": string,
-    "code": string,
-    "retryable": boolean,
-    "details": string | null
-  }
-}
+"""Summarize worker.
+Reads JSON from stdin and writes JSON to stdout.
 """
 
-import json
 import hashlib
-import re
-import os
-import sys
-from typing import Any, Optional
+from typing import Any
 
 from common.errors import ErrorCode, LLMError, ValidationError, WorkerError, classify_exception
 from common.logging import get_logger
+from common.llm_adapters import generate as llm_generate
+from summarize_extract import (
+    PARSER_VERSION,
+    collect_scale_labels,
+    extract_decision_code,
+    extract_decision_code_from_text,
+    extract_leading_decision_code,
+    extract_leading_text_label_decision,
+    extract_leading_text_label_decision_relaxed,
+    extract_text_label_decision,
+    extract_text_label_decision_relaxed,
+)
+from summarize_llm import (
+    DEFAULT_SUMMARY_MODEL,
+    LLM_FALLBACK_MODEL,
+    build_llm_decision_prompt,
+    classify_decision_with_llm,
+)
+from summarize_text import build_response_text, response_excerpt
 
 log = get_logger("summarize")
 
-# Pattern to find "Rating: X" format (preferred, structured)
-STRUCTURED_RATING_PATTERN = re.compile(r"Rating:\s*([1-9]\d*)", re.IGNORECASE)
-
-# Additional structured formats commonly returned by models.
-STRUCTURED_DECISION_PATTERNS = [
-    re.compile(r"\b(?:decision(?:\s*code)?|answer|response)\s*(?:is|=|:)?\s*([1-9]\d*)\b", re.IGNORECASE),
-    re.compile(r"\b(?:my\s+)?judg(?:e)?ment(?:\s+on\s+the\s+scale)?\s*(?:(?:is)\s*[:=]?|[:=])?\s*(?:a\s*)?([1-9]\d*)\b", re.IGNORECASE),
-    re.compile(r"\bmy\s+(?:rating|answer|response|choice)\s*(?:is|=|:)?\s*([1-9]\d*)\b", re.IGNORECASE),
-    re.compile(r"\b(?:my|overall|final)\s+score\s*(?:is|=|:)?\s*([1-9]\d*)\b", re.IGNORECASE),
-    re.compile(r"\bi\s*(?:choose|chose|pick|picked|select|selected|rate|rated|would choose|would pick|would select)\s*(?:option\s*)?([1-9]\d*)\b", re.IGNORECASE),
-    re.compile(r"\b(?:i\s+would\s+)?lean\s+toward\s*(?:option\s*)?([1-9]\d*)\b", re.IGNORECASE),
-    re.compile(r"\bi(?:['’]d|\s+would)?\s+rate\b.{0,40}?([1-9]\d*)\b", re.IGNORECASE),
-    re.compile(r"\bi(?:['’]d|\s+would)?\s+(?:give|assign)\b.{0,30}?([1-9]\d*)\b", re.IGNORECASE),
-    re.compile(r"\b([1-9]\d*)\s*(?:is|=|:)?\s*(?:my|the)\s*(?:rating|answer|response|choice)\b", re.IGNORECASE),
-    re.compile(r"\b([1-9]\d*)\s*[—–-]\s*(?:strongly|somewhat|neutral|not like|like me|support)\b", re.IGNORECASE),
-    re.compile(r"^\s*([1-9]\d*)\s*(?:reason\s*:)", re.IGNORECASE | re.MULTILINE),
-    re.compile(r"^\s*([1-9]\d*)\s*(?:\r?\n)+\s*\*{0,2}reason(?:ing)?\*{0,2}\s*:", re.IGNORECASE | re.MULTILINE),
-    re.compile(r"^\s*([1-9]\d*)\s*[—-]\s*", re.IGNORECASE | re.MULTILINE),
-]
-
-# Fallback pattern to find positive integer ratings - less reliable
-FALLBACK_RATING_PATTERN = re.compile(r"\b([1-9]\d*)\b")
-RANGE_PATTERN = re.compile(r"([1-9]\d*)\s*(?:-|–|—|to)\s*([1-9]\d*)", re.IGNORECASE)
-# Word-count parentheticals appended by some models, e.g. "(5 words)" or "(152 words)".
-# Must be stripped before fallback numeric scanning to avoid false code matches.
-WORD_COUNT_SUFFIX_PATTERN = re.compile(r"\(\s*\d+\s+words?\s*\)", re.IGNORECASE)
-AMBIGUOUS_SUFFIX_PATTERN = re.compile(r"^\s*(?:and|or|/|,)\s*([1-9]\d*)\b", re.IGNORECASE)
-LEADING_BARE_NUMBER_PATTERN = re.compile(r"^\s*([1-9]\d*)\s*(?:[.!?])?\s*$", re.IGNORECASE)
-
-# Heuristics for detecting responses that are about the user ("you/your")
-# rather than a self-rating from the target AI.
-USER_DIRECTED_PATTERN = re.compile(
-    r"\b("
-    r"would you|do you|what do you think|would you like|"
-    r"you should|you(?:'d| would)\s+likely|which option"
-    r")\b",
-    re.IGNORECASE,
-)
-SELF_RATING_PATTERN = re.compile(r"\b(i|i'm|i’d|i'd|i would|my|for me|personally)\b", re.IGNORECASE)
-REFUSAL_PATTERN = re.compile(
-    r"\b("
-    r"must decline|decline to (?:engage|respond)|"
-    r"can't comply|cannot comply|won't comply|"
-    r"jailbreak|circumvent (?:my )?(?:safety|guidelines)|"
-    r"adhere to my safety instructions|"
-    r"i(?: am|'m) sorry[, ]+but"
-    r")\b",
-    re.IGNORECASE,
-)
-LEADING_DECISION_PREFIX_PATTERN = re.compile(
-    r"^\s*(?:"
-    r"(?:my\s+)?(?:final\s+|overall\s+)?(?:judg(?:e)?ment|answer|response|decision|choice|rating|score)"
-    r"(?:\s+on\s+the\s+scale)?\s*(?:(?:is)\s*[:=]?|[:=])?\s*"
-    r"|level\s+of\s+support\s*[:=]?\s*"
-    r"|in\s+short\s*[:,-]?\s*"
-    r"|overall\s*[:,-]?\s*"
-    r")",
-    re.IGNORECASE,
-)
-
-# Default summary model if none specified
-DEFAULT_SUMMARY_MODEL = "anthropic:claude-sonnet-4-20250514"
-LLM_FALLBACK_MODEL = "xai:grok-4-1-fast-reasoning"
-DEFAULT_SUMMARIZE_PARSER_VERSION = "paired-v2"
+# Compatibility hook for tests that patch `summarize.generate`.
+generate = llm_generate
 
 # Keep batch size aligned with the API-side PgBoss worker batch size ceiling.
 MAX_SUMMARIZE_BATCH_SIZE = 500
 
 
-def load_parser_version() -> str:
-    raw_version = os.getenv("SUMMARIZE_PARSER_VERSION", DEFAULT_SUMMARIZE_PARSER_VERSION)
-    normalized_version = raw_version.strip()
-    return normalized_version if normalized_version else DEFAULT_SUMMARIZE_PARSER_VERSION
-
-
-PARSER_VERSION = load_parser_version()
+def _validation_error(message: str, details: str | None = None) -> dict[str, Any]:
+    return {
+        "message": message,
+        "code": ErrorCode.VALIDATION_ERROR.value,
+        "retryable": False,
+        "details": details,
+    }
 
 
 def validate_input(data: dict[str, Any]) -> None:
@@ -160,455 +71,11 @@ def is_batch_envelope(data: dict[str, Any]) -> bool:
     return has_batch_field and not has_single_fields
 
 
-def normalize_for_match(text: str) -> str:
-    sanitized = text.replace("**", " ").replace("__", " ").replace("`", " ")
-    sanitized = re.sub(r"[^a-z0-9]+", " ", sanitized.lower())
-    return re.sub(r"\s+", " ", sanitized).strip()
-
-
-_FILLER_WORDS = frozenset({"their", "the", "a", "an"})
-
-
-def normalize_for_relaxed_match(text: str) -> str:
-    base = normalize_for_match(text)
-    return " ".join(w for w in base.split() if w not in _FILLER_WORDS)
-
-
-def build_response_text(transcript_content: dict[str, Any]) -> str:
-    turns = transcript_content.get("turns", [])
-    responses: list[str] = []
-    for turn in turns:
-        response = turn.get("targetResponse", "")
-        if isinstance(response, str) and response:
-            responses.append(response)
-    return "\n".join(responses).strip()
-
-
-def response_excerpt(text: str, limit: int = 280) -> str:
-    compact = re.sub(r"\s+", " ", text).strip()
-    return compact[:limit]
-
-
-def response_segments(text: str) -> list[str]:
-    segments: list[str] = []
-    for block in re.split(r"[\n\r]+", text):
-        stripped = block.strip()
-        if not stripped:
-            continue
-        segments.extend(
-            sentence.strip()
-            for sentence in re.split(r"(?<=[.!?])\s+", stripped)
-            if sentence.strip()
-        )
-    return segments
-
-
-def strip_leading_decision_prefix(text: str) -> str:
-    if not text:
-        return ""
-
-    stripped = text.replace("**", "").replace("__", "").replace("`", "").strip()
-    previous = None
-    while stripped and stripped != previous:
-        previous = stripped
-        stripped = LEADING_DECISION_PREFIX_PATTERN.sub("", stripped, count=1).strip()
-    return stripped
-
-
-def leading_response_candidates(text: str) -> list[tuple[str, bool]]:
-    if not text:
-        return []
-
-    candidates: list[tuple[str, bool]] = []
-    lines = [line.strip() for line in re.split(r"[\n\r]+", text) if line.strip()]
-    segments = response_segments(text)
-
-    for candidate in [
-        lines[0] if lines else "",
-        segments[0] if segments else "",
-    ]:
-        if not candidate:
-            continue
-        stripped = strip_leading_decision_prefix(candidate)
-        for value, used_prefix_stripping in [
-            (candidate, False),
-            (stripped, stripped != candidate),
-        ]:
-            if value and not any(existing == value for existing, _ in candidates):
-                candidates.append((value, used_prefix_stripping))
-
-    return candidates
-
-
-def collect_scale_labels(transcript_content: dict[str, Any]) -> list[dict[str, str]]:
-    turns = transcript_content.get("turns", [])
-    for turn in turns:
-        probe_prompt = turn.get("probePrompt")
-        if not isinstance(probe_prompt, str) or probe_prompt.strip() == "":
-            continue
-
-        numbered_labels: list[dict[str, str]] = []
-        bullet_labels: list[str] = []
-
-        for raw_line in probe_prompt.splitlines():
-            line = raw_line.strip()
-            if line == "":
-                continue
-
-            numbered_match = re.match(r"^(?P<code>[1-9]\d*)\s*-\s*(?P<label>.+)$", line)
-            if numbered_match:
-                numbered_labels.append(
-                    {
-                        "code": numbered_match.group("code"),
-                        "label": numbered_match.group("label").strip(),
-                    }
-                )
-                continue
-
-            bullet_match = re.match(r"^-\s+(?P<label>.+)$", line)
-            if bullet_match:
-                bullet_labels.append(bullet_match.group("label").strip())
-
-        if len(numbered_labels) >= 5:
-            return numbered_labels
-        if len(bullet_labels) == 5:
-            return [
-                {"code": code, "label": label}
-                for code, label in zip(["5", "4", "3", "2", "1"], bullet_labels)
-            ]
-
-    return []
-
-
-def extract_text_label_decision(text: str, scale_labels: list[dict[str, str]]) -> tuple[Optional[str], Optional[str]]:
-    if not text or not scale_labels:
-        return None, None
-
-    segments = response_segments(text)
-    if not segments:
-        return None, None
-
-    normalized_labels = [
-        {
-            "code": entry.get("code", ""),
-            "label": entry.get("label", ""),
-            "normalized": normalize_for_match(entry.get("label", "")),
-        }
-        for entry in scale_labels
-        if entry.get("label")
-    ]
-
-    for segment in segments:
-        normalized_segment = normalize_for_match(segment)
-        if normalized_segment == "":
-            continue
-
-        prefix_matches = [
-            entry
-            for entry in normalized_labels
-            if entry["normalized"]
-            and (
-                normalized_segment == entry["normalized"]
-                or normalized_segment.startswith(entry["normalized"] + " ")
-            )
-        ]
-
-        unique_prefix_matches = list({entry["code"]: entry for entry in prefix_matches}.values())
-        if len(unique_prefix_matches) == 1:
-            match = unique_prefix_matches[0]
-            return match["code"], match["label"]
-        if len(unique_prefix_matches) > 1:
-            return None, None
-
-    return None, None
-
-
-def extract_text_label_decision_relaxed(
-    text: str, scale_labels: list[dict[str, str]]
-) -> tuple[Optional[str], Optional[str]]:
-    """Same as extract_text_label_decision but strips filler words before comparing."""
-    if not text or not scale_labels:
-        return None, None
-
-    segments = response_segments(text)
-    if not segments:
-        return None, None
-
-    normalized_labels = [
-        {
-            "code": entry.get("code", ""),
-            "label": entry.get("label", ""),
-            "relaxed": normalize_for_relaxed_match(entry.get("label", "")),
-        }
-        for entry in scale_labels
-        if entry.get("label")
-    ]
-
-    for segment in segments:
-        relaxed_segment = normalize_for_relaxed_match(segment)
-        if relaxed_segment == "":
-            continue
-
-        prefix_matches = [
-            entry
-            for entry in normalized_labels
-            if entry["relaxed"]
-            and (
-                relaxed_segment == entry["relaxed"]
-                or relaxed_segment.startswith(entry["relaxed"] + " ")
-            )
-        ]
-
-        unique_prefix_matches = list({entry["code"]: entry for entry in prefix_matches}.values())
-        if len(unique_prefix_matches) == 1:
-            match = unique_prefix_matches[0]
-            return match["code"], match["label"]
-        if len(unique_prefix_matches) > 1:
-            return None, None
-
-    return None, None
-
-
-def extract_leading_text_label_decision_relaxed(
-    text: str, scale_labels: list[dict[str, str]]
-) -> tuple[Optional[str], Optional[str], Optional[str]]:
-    """Same as extract_leading_text_label_decision but uses relaxed matching."""
-    for candidate, used_prefix_stripping in leading_response_candidates(text):
-        decision_code, matched_label = extract_text_label_decision_relaxed(candidate, scale_labels)
-        if decision_code is not None:
-            parse_path = "text_label_relaxed_leading" if used_prefix_stripping else "text_label_relaxed"
-            return decision_code, matched_label, parse_path
-    return None, None, None
-
-
-def extract_leading_decision_code(text: str) -> Optional[str]:
-    for candidate, _used_prefix_stripping in leading_response_candidates(text):
-        decision_code = extract_explicit_leading_decision_code(candidate)
-        if decision_code is not None:
-            return decision_code
-    return None
-
-
-def extract_leading_text_label_decision(
-    text: str, scale_labels: list[dict[str, str]]
-) -> tuple[Optional[str], Optional[str], Optional[str]]:
-    for candidate, used_prefix_stripping in leading_response_candidates(text):
-        decision_code, matched_label = extract_text_label_decision(candidate, scale_labels)
-        if decision_code is not None:
-            parse_path = "text_label_leading" if used_prefix_stripping else "text_label_exact"
-            return decision_code, matched_label, parse_path
-    return None, None, None
-
-
-def extract_decision_code_from_text(text: str) -> Optional[str]:
-    """
-    Extract numeric decision code (positive integer) from text.
-
-    First looks for structured "Rating: X" format (most reliable).
-    Falls back to finding first standalone positive integer (less reliable).
-
-    Returns None if no rating found.
-    """
-    if not text:
-        return None
-
-    # Strip lightweight markdown markers that often surround numeric answers.
-    sanitized_markdown_text = text.replace("**", "").replace("__", "").replace("`", "")
-
-    # First, try structured "Rating: X" format (most reliable)
-    structured_match = STRUCTURED_RATING_PATTERN.search(sanitized_markdown_text)
-    if structured_match:
-        suffix = sanitized_markdown_text[structured_match.end():structured_match.end() + 24]
-        if AMBIGUOUS_SUFFIX_PATTERN.search(suffix):
-            return None
-        return structured_match.group(1)
-
-    # Next, look for common explicit decision formats.
-    for pattern in STRUCTURED_DECISION_PATTERNS:
-        matches = []
-        for match in pattern.finditer(sanitized_markdown_text):
-            # If immediate continuation suggests multiple candidate codes
-            # (e.g., "I choose 3 and 4"), treat as ambiguous.
-            suffix = sanitized_markdown_text[match.end():match.end() + 24]
-            if AMBIGUOUS_SUFFIX_PATTERN.search(suffix):
-                return None
-            matches.append(match.group(1))
-        if not matches:
-            continue
-        unique_values = list(dict.fromkeys(matches))
-        if len(unique_values) == 1:
-            return unique_values[0]
-        # If an explicit format provides conflicting values, treat as ambiguous.
-        return None
-
-    # Remove numeric ranges (e.g., "1-6", "1 to 6") before fallback scanning.
-    sanitized_text = RANGE_PATTERN.sub(" ", sanitized_markdown_text)
-    # Strip word-count parentheticals like "(5 words)" before fallback scanning.
-    # Some models append these as metadata; a coincidentally valid code would cause
-    # a false positive that the out-of-range guard cannot catch.
-    sanitized_text = WORD_COUNT_SUFFIX_PATTERN.sub(" ", sanitized_text)
-
-    fallback_matches = [m.group(1) for m in FALLBACK_RATING_PATTERN.finditer(sanitized_text)]
-    if not fallback_matches:
-        if REFUSAL_PATTERN.search(text):
-            return "refusal"
-        return None
-
-    # If the model enumerates multiple options (e.g., "1..6"), it did not choose one.
-    unique_values = list(dict.fromkeys(fallback_matches))
-    if len(unique_values) > 1:
-        return None
-
-    # If response is clearly directed at the user and lacks self-rating language,
-    # treat as no valid decision choice.
-    lower_text = text.lower()
-    has_user_directed = USER_DIRECTED_PATTERN.search(lower_text) is not None
-    has_self_rating = SELF_RATING_PATTERN.search(lower_text) is not None
-    if has_user_directed and not has_self_rating:
-        return None
-
-    return unique_values[0]
-
-    return None
-
-
-def extract_explicit_leading_decision_code(text: str) -> Optional[str]:
-    """
-    Extract a decision from a leading candidate only when the candidate contains
-    an explicit decision signal.
-
-    This intentionally avoids the broad fallback number scan used for whole-response
-    parsing, because opening lines often contain contextual numbers that are not
-    the final decision.
-    """
-    if not text:
-        return None
-
-    sanitized_markdown_text = text.replace("**", "").replace("__", "").replace("`", "")
-
-    structured_match = STRUCTURED_RATING_PATTERN.search(sanitized_markdown_text)
-    if structured_match:
-        suffix = sanitized_markdown_text[structured_match.end():structured_match.end() + 24]
-        if AMBIGUOUS_SUFFIX_PATTERN.search(suffix):
-            return None
-        return structured_match.group(1)
-
-    for pattern in STRUCTURED_DECISION_PATTERNS:
-        matches = []
-        for match in pattern.finditer(sanitized_markdown_text):
-            suffix = sanitized_markdown_text[match.end():match.end() + 24]
-            if AMBIGUOUS_SUFFIX_PATTERN.search(suffix):
-                return None
-            matches.append(match.group(1))
-        if not matches:
-            continue
-        unique_values = list(dict.fromkeys(matches))
-        if len(unique_values) == 1:
-            return unique_values[0]
-        return None
-
-    bare_number_match = LEADING_BARE_NUMBER_PATTERN.match(sanitized_markdown_text)
-    if bare_number_match:
-        return bare_number_match.group(1)
-
-    if REFUSAL_PATTERN.search(text):
-        return "refusal"
-
-    return None
-
-
-def extract_decision_code(transcript_content: dict[str, Any]) -> str:
-    """
-    Extract numeric decision code (positive integer) from transcript.
-
-    Looks for "Rating: X" format first (structured response from probe).
-    Falls back to finding first positive integer in the response.
-
-    Returns "other" if no clear rating/refusal found.
-    """
-    turns = transcript_content.get("turns", [])
-
-    # Combine all target responses
-    full_text = ""
-    for turn in turns:
-        response = turn.get("targetResponse", "")
-        if response:
-            full_text += response + "\n"
-
-    result = extract_decision_code_from_text(full_text)
-    return result if result else "other"
-
-
-def build_llm_decision_prompt(
-    transcript_content: dict[str, Any], scale_labels: Optional[list[dict[str, str]]] = None
-) -> str:
-    """Build strict prompt for fallback LLM decision classification."""
-    transcript_text = build_response_text(transcript_content)
-    scale_section = ""
-    if scale_labels:
-        formatted_labels = "\n".join(
-            f"- {entry['code']}: {entry['label']}" for entry in scale_labels
-        )
-        scale_section = f"Available scale labels:\n{formatted_labels}\n\n"
-
-    return (
-        "You are extracting a single final decision code from an AI response.\n"
-        "Return exactly one token:\n"
-        "- a positive integer (e.g. 1, 2, 6)\n"
-        "- refusal\n"
-        "- other\n"
-        "No explanation, no punctuation, no extra words.\n\n"
-        f"{scale_section}"
-        "Transcript:\n"
-        f"{transcript_text}\n"
-        "\nAnswer:"
-    )
-
-
-def classify_decision_with_llm(
-    transcript_content: dict[str, Any], scale_labels: Optional[list[dict[str, str]]] = None
-) -> str:
-    """
-    Use fallback LLM to classify an unresolved decision.
-
-    Returns decision code string, "refusal", or "other".
-    """
-    # Lazy import: deferred until actually needed so that the heavy LLM SDK
-    # dependencies are not loaded for the 99%+ of jobs that resolve deterministically.
-    from common.llm_adapters import generate  # noqa: PLC0415
-
-    prompt = build_llm_decision_prompt(transcript_content, scale_labels)
-    messages = [{"role": "user", "content": prompt}]
-
-    try:
-        response = generate(
-            LLM_FALLBACK_MODEL,
-            messages,
-            temperature=0.0,
-            max_tokens=20,
-        )
-        normalized = response.content.strip().splitlines()[0].strip().lower()
-        if normalized == "":
-            return "other"
-        if normalized == "refusal":
-            return "refusal"
-        if normalized == "other":
-            return "other"
-        numeric_match = re.search(r"\b([1-9]\d*)\b", normalized)
-        if numeric_match:
-            return numeric_match.group(1)
-        return "other"
-    except (WorkerError, LLMError) as err:
-        log.error("Fallback LLM decision classification failed", err=err)
-        return "other"
-    except Exception as err:
-        log.error("Unexpected error in fallback LLM decision classification", err=err)
-        return "other"
-
-
 def extract_decision_result(transcript_content: dict[str, Any]) -> dict[str, Any]:
     response_text = build_response_text(transcript_content)
-    response_hash = hashlib.sha256(response_text.encode("utf-8")).hexdigest() if response_text else None
+    response_hash = (
+        hashlib.sha256(response_text.encode("utf-8")).hexdigest() if response_text else None
+    )
     scale_labels = collect_scale_labels(transcript_content)
 
     leading_decision_code = extract_leading_decision_code(response_text)
@@ -627,7 +94,9 @@ def extract_decision_result(transcript_content: dict[str, Any]) -> dict[str, Any
             decision_code = "other"
 
     if decision_code == "other" and scale_labels:
-        text_label_code, matched_label, leading_text_label_path = extract_leading_text_label_decision(response_text, scale_labels)
+        text_label_code, matched_label, leading_text_label_path = (
+            extract_leading_text_label_decision(response_text, scale_labels)
+        )
         if text_label_code is not None:
             decision_code = text_label_code
             parse_path = leading_text_label_path or "text_label_exact"
@@ -640,17 +109,24 @@ def extract_decision_result(transcript_content: dict[str, Any]) -> dict[str, Any
                 # Relaxed matching: strip filler words (their/the/a/an) before comparing.
                 # Catches models that paraphrase slightly (e.g. "recognition of expertise"
                 # instead of "recognition of their expertise").
-                relaxed_code, matched_label, relaxed_path = extract_leading_text_label_decision_relaxed(response_text, scale_labels)
+                relaxed_code, matched_label, relaxed_path = (
+                    extract_leading_text_label_decision_relaxed(response_text, scale_labels)
+                )
                 if relaxed_code is not None:
                     decision_code = relaxed_code
                     parse_path = relaxed_path or "text_label_relaxed"
                 else:
-                    relaxed_code, matched_label = extract_text_label_decision_relaxed(response_text, scale_labels)
+                    relaxed_code, matched_label = extract_text_label_decision_relaxed(
+                        response_text, scale_labels
+                    )
                     if relaxed_code is not None:
                         decision_code = relaxed_code
                         parse_path = "text_label_relaxed"
                     else:
-                        llm_decision_code = classify_decision_with_llm(transcript_content, scale_labels)
+                        llm_decision_code = classify_decision_with_llm(
+                            transcript_content,
+                            scale_labels,
+                        )
                         if llm_decision_code != "other":
                             decision_code = llm_decision_code
                             decision_source = "llm"
@@ -753,6 +229,8 @@ def run_summarize(data: dict[str, Any]) -> dict[str, Any]:
         return {
             "success": True,
             "summary": {
+                "decisionCode": decision_code,
+                "decisionSource": decision_source,
                 "decisionText": decision_text,
                 "decisionMetadata": decision_metadata,
             },
@@ -772,190 +250,7 @@ def run_summarize(data: dict[str, Any]) -> dict[str, Any]:
             "error": worker_err.to_dict(),
         }
 
-
-def run_summarize_batch(data: dict[str, Any]) -> dict[str, Any]:
-    """Execute summarization for a batch of transcripts."""
-    transcripts = data.get("transcripts", [])
-    if not isinstance(transcripts, list):
-        return {
-            "success": False,
-            "error": {
-                "message": "transcripts must be an array",
-                "code": ErrorCode.VALIDATION_ERROR.value,
-                "retryable": False,
-            },
-        }
-    if len(transcripts) > MAX_SUMMARIZE_BATCH_SIZE:
-        return {
-            "success": False,
-            "error": {
-                "message": f"Batch size exceeds maximum of {MAX_SUMMARIZE_BATCH_SIZE}",
-                "code": ErrorCode.VALIDATION_ERROR.value,
-                "retryable": False,
-            },
-        }
-
-    summaries: list[dict[str, Any]] = []
-    failure_count = 0
-    retryable_failure = False
-    for batch_index, item in enumerate(transcripts):
-        if not isinstance(item, dict):
-            failure_count += 1
-            summaries.append(
-                {
-                    "transcriptId": None,
-                    "batchIndex": batch_index,
-                    "success": False,
-                    "error": {
-                        "message": "Transcript item must be an object",
-                        "code": ErrorCode.VALIDATION_ERROR.value,
-                        "retryable": False,
-                    },
-                }
-            )
-            continue
-
-        transcript_id = item.get("transcriptId")
-        try:
-            validate_input(item)
-            result = run_summarize(item)
-        except ValidationError as err:
-                result = {
-                    "success": False,
-                    "error": err.to_dict(),
-                }
-        except Exception as err:
-            worker_err = classify_exception(err)
-            result = {
-                "success": False,
-                "error": worker_err.to_dict(),
-            }
-        if not result.get("success", False):
-            failure_count += 1
-            error = result.get("error")
-            if isinstance(error, dict) and error.get("retryable") is True:
-                retryable_failure = True
-        summaries.append(
-            {
-                "transcriptId": transcript_id,
-                "batchIndex": batch_index,
-                **result,
-            }
-        )
-
-    if failure_count > 0:
-        return {
-            "success": False,
-            "error": {
-                "message": f"{failure_count} transcript(s) failed",
-                "code": "BATCH_PARTIAL_FAILURE",
-                "retryable": retryable_failure,
-                "details": {
-                    "failedCount": failure_count,
-                    "totalCount": len(transcripts),
-                    "maxBatchSize": MAX_SUMMARIZE_BATCH_SIZE,
-                },
-            },
-            "summaries": summaries,
-        }
-
-    return {
-        "success": True,
-        "summaries": summaries,
-    }
-
-
-def main() -> None:
-    """Main entry point - read from stdin, write to stdout."""
-    try:
-        # Read JSON input from stdin
-        input_data = sys.stdin.read()
-        if not input_data.strip():
-            result = {
-                "success": False,
-                "error": {
-                    "message": "No input provided",
-                    "code": ErrorCode.VALIDATION_ERROR.value,
-                    "retryable": False,
-                },
-            }
-            print(json.dumps(result))
-            return
-
-        try:
-            data = json.loads(input_data)
-        except json.JSONDecodeError as err:
-            result = {
-                "success": False,
-                "error": {
-                    "message": f"Invalid JSON input: {err}",
-                    "code": ErrorCode.VALIDATION_ERROR.value,
-                    "retryable": False,
-                },
-            }
-            print(json.dumps(result))
-            return
-
-        if isinstance(data, dict):
-            if "transcripts" in data:
-                if not is_batch_envelope(data):
-                    result = {
-                        "success": False,
-                        "error": {
-                            "message": "Batch envelopes cannot include transcriptId or transcriptContent",
-                            "code": ErrorCode.VALIDATION_ERROR.value,
-                            "retryable": False,
-                        },
-                    }
-                    print(json.dumps(result))
-                    return
-
-                result = run_summarize_batch(data)
-            else:
-                # Validate input
-                try:
-                    validate_input(data)
-                except ValidationError as err:
-                    result = {
-                        "success": False,
-                        "error": err.to_dict(),
-                    }
-                    print(json.dumps(result))
-                    return
-
-                # Run summarization
-                result = run_summarize(data)
-        else:
-            # Validate input
-            try:
-                validate_input(data)
-            except ValidationError as err:
-                result = {
-                    "success": False,
-                    "error": err.to_dict(),
-                }
-                print(json.dumps(result))
-                return
-            # Run summarization
-            result = run_summarize(data)
-
-        # Output result
-        print(json.dumps(result))
-
-    except Exception as err:
-        # Catch-all for unexpected errors
-        log.error("Unexpected error in summarize worker", err=err)
-        result = {
-            "success": False,
-            "error": {
-                "message": str(err),
-                "code": ErrorCode.UNKNOWN.value,
-                "retryable": True,
-                "details": type(err).__name__,
-            },
-        }
-        print(json.dumps(result))
-
-
 if __name__ == "__main__":
-    main()
+    import summarize_batch
+
+    summarize_batch.main()
